chore: Add Dockerfile for containerizing the application
This commit introduces a Dockerfile to package the Java 11 runtime environment and the application JAR file. The Dockerfile uses the official openjdk:11-jre-slim base image, copies the application JAR file into the container, sets the working directory, and specifies the command to run the application upon container startup.

Details:
- Base image: openjdk:11-jre-slim
- Application JAR path: ./target/gsmarena-solvd-1.0-SNAPSHOT.jar
- Container working directory: /app
- Run command: java -jar gsmarena-solvd.jar

This Dockerfile is designed to simplify the deployment and execution of the application within a containerized environment.